implemented dialogue paging
that took more time than I hoped for, but this was an important one to finish...

dialogue text is now an array fo pages, if array has more than one item (or page) no replies are shown until last page is displayed and if you click on dialogue text, it displays next page.

what took time was setting the right flags in the right place, especially since  I was already tired.