Finished conditional dialogue, added demo dialogue quest
This was supposed to be the final addition, but I realized I want to be able to change more than one dialogue per reply. Also, instead of the game_progression func, I could change game_vars dynamically with json in pick_response func.

"affected": "npc1"
"branch": "e"

will save a lot of code in the long run.

Not a huge priority though... this "works"